backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
Django==4.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
idna==3.10
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pyheif==0.8.0
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
